pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "fastapi>=0.118.0",
    "numpy>=2.3.3",
    "ortools>=9.14.6206",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
    "xgboost>=3.0.5",
]
